Fixes
Fixed the flag preventing a battle with Malignius Mortifier from triggering.

Removed the value from the rotten potato and pulse core items to hopefully prevent these from being able to be sold to shops.

Fixed an issue with a haystack on Route 12 that clipped through the player's head.

Attempted to hide the Duke from being visible on the top floor of Lumbridge Castle before the player has interacted with him